% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20079.R
\name{FCM}
\alias{FCM}
\title{FCM}
\usage{
FCM(dat, k)
}
\arguments{
\item{dat}{training set(matrix)}

\item{k}{Number of clusters(numeric)}
}
\value{
a list (u,v,cluster_id)
}
\description{
Fuzzy cluster analysis of data
}
\examples{
\dontrun{
data(iris)
dat <- iris[,2:4]
result <- FCM(dat,3)
plot(iris$Sepal.Width,iris$Petal.Width,col=result$cluster_id,main="Visualization results",pch=19)
}
}
