% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20079.R
\name{Lagrange}
\alias{Lagrange}
\title{Lagrange interpolation .}
\usage{
Lagrange(x, X, Y)
}
\arguments{
\item{x}{The x value corresponding to the missing value y (numeric)}

\item{X}{The x value vector corresponding to the Y vector(numeric)}

\item{Y}{All known values of y (numeric)}
}
\value{
A vector (y)
}
\description{
Approximate unknown function with Lagrange interpolation
}
\examples{
\dontrun{
X <- c(-2.15,-1.00,0.01,1.02,2.03,3.25)
Y <- c(17.03,7.24,1.05,2.03,17.06,23.05)
f <- function(x)
 Lagrange(x,X,Y)
curve(f(x),-2.15,3.25)
}
}
