\name{StatComp20079}
\alias{StatComp20079}
\docType{package}
\title{
  This package provides lagrangian interpolation function and fuzzy cluster analysis.
}
\description{
  Approximate unknown function with Lagrange interpolation and fuzzy cluster analysis of data.
}
\author{
Yueping Zhang

Maintainer: Yueping Zhang <zhangyuepingwa@163.com>
}
\keyword{ package }
